setup/requirements.txt
matplotlib
mplfinance
numpy
pandas
python-dotenv
scipy
tqdm
yfinance